requirements.txt
pycryptodome==3.23.0
Flask==3.1.2
